requirements.txt
requests
pyodbc
python-dotenv
openai 